Get hltb command working
Also update itad
