--- src/hltb.js
@@ -1,4 +1,0 @@
-const hltb = require('howlongtobeat')
-
-// Initialize HLTB Service
-var hltbService = new hltb.HowLongToBeatService()
